Tag and release 4.12
git-svn-id: https://plugins.svn.wordpress.org/subscribe2/tags/4.12@79119 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ subscribe2.php
@@ -263,7 +263,8 @@ class s2class {
 				// To send HTML mail, the Content-Type header must be set
 				$headers .= "MIME-Version: 1.0\n";
 				$headers .= "Content-Type: " . get_bloginfo('html_type') . "; charset=\"". get_bloginfo('charset') . "\"\n";
-				$mailtext = "<html><head><title>" . $subject . "</title></head><body>" . $message . "</body></html>";
+				$mailtext = "<html><head><title>" . $subject . "</title><link rel=\"stylesheet\" href=\"" . get_bloginfo('stylesheet_url') . "\"
+type=\"text/css\" media=\"screen\" /></head><body>" . $message . "</body></html>";
 		} else {
 				$headers .= "MIME-Version: 1.0\n";
 				$headers .= "Content-Type: text/plain; charset=\"". get_bloginfo('charset') . "\"\n";
@@ -671,9 +672,6 @@ class s2class {
 	*/
 	function remind($emails = '') {
 		if ('' == $emails) { return false; }
-
-		$admin = $this->get_userdata();
-		$this->myname = $admin->display_name;
 		
 		$recipients = explode(",", $emails);
 		if (!is_array($recipients)) { $recipients = (array)$recipients; }
@@ -720,19 +718,34 @@ class s2class {
 			return $this->no_such_email;
 		}
 
+		// get current status of email so messages are only sent once per emailed link
+		$current = $this->is_public($this->email);
+
 		if ('1' == $action) {
 			// make this subscription active
-			$this->activate();
 			$this->message = $this->added;
-			$subject = '[' . get_option('blogname') . '] ' . __('New subscriber', 'subscribe2');
-			$message = $this->email . " " . __('subscribed to email notifications!', 'subscribe2');
-			$recipients = $wpdb->get_col("SELECT DISTINCT(user_email) FROM $wpdb->users INNER JOIN $wpdb->usermeta ON $wpdb->users.ID = $wpdb->usermeta.user_id WHERE $wpdb->usermeta.meta_key='" . $wpdb->prefix . "user_level' AND $wpdb->usermeta.meta_value='10'");
-			$this->mail($recipients, $subject, $message);
+			if ('1' != $current) {
+				$this->activate();
+				if ( ($this->subscribe2_options['admin_email'] == 'subs') || ($this->subscribe2_options['admin_email'] == 'both') ) {
+					$subject = '[' . get_option('blogname') . '] ' . __('New subscription', 'subscribe2');
+					$message = $this->email . " " . __('subscribed to email notifications!', 'subscribe2');
+					$recipients = $wpdb->get_col("SELECT DISTINCT(user_email) FROM $wpdb->users INNER JOIN $wpdb->usermeta ON $wpdb->users.ID = $wpdb->usermeta.user_id WHERE $wpdb->usermeta.meta_key='" . $wpdb->prefix . "user_level' AND $wpdb->usermeta.meta_value='10'");
+					$this->mail($recipients, $subject, $message);
+				}
+			}
 			$this->filtered = 1;
 		} elseif ('0' == $action) {
 			// remove this subscriber
-			$this->delete();
 			$this->message = $this->deleted;
+			if ('0' != $current) {
+				$this->delete();
+				if ( ($this->subscribe2_options['admin_email'] == 'unsubs') || ($this->subscribe2_options['admin_email'] == 'both') ) {
+					$subject = '[' . get_option('blogname') . '] ' . __('New Unsubscription', 'subscribe2');
+					$message = $this->email . " " . __('unsubscribed from email notifications!', 'subscribe2');
+					$recipients = $wpdb->get_col("SELECT DISTINCT(user_email) FROM $wpdb->users INNER JOIN $wpdb->usermeta ON $wpdb->users.ID = $wpdb->usermeta.user_id WHERE $wpdb->usermeta.meta_key='" . $wpdb->prefix . "user_level' AND $wpdb->usermeta.meta_value='10'");
+					$this->mail($recipients, $subject, $message);
+				}
+			}
 			$this->filtered = 1;
 		}
 
@@ -1235,7 +1248,7 @@ class s2class {
 			// Displays a page number strip - adapted from code in Akismet
 			$args['what'] = $what;
 			$total_subscribers = count($subscribers);
-			$total_pages = ceil($total_subscribers / 25);
+			$total_pages = ceil($total_subscribers / $this->subscribe2_options['entries']);
 			$strip = '';
 			if ( $page > 1 ) {
 				$args['s2page'] = $page - 1;
@@ -1257,19 +1270,20 @@ class s2class {
 					}
 				}
 			}
-			if ( ( $page ) * 25 < $total_subscribers ) {
+			if ( ( $page ) * $this->subscribe2_options['entries'] < $total_subscribers ) {
 				$args['s2page'] = $page + 1;
 				$strip .=  "<a class=\"next\" href=\"" . clean_url(add_query_arg($args)) . "\">". __('Next Page', 'subscribe2') . " &raquo;</a>\n";
 			}
 		}
 
 		// show our form
-		echo "<form method=\"post\" action=\"\">\r\n";
 		echo "<div class=\"wrap\">";
 		echo "<h2>" . __('Manage Subscribers', 'subscribe2') . "</h2>\r\n";
+		echo "<form method=\"post\" action=\"\">\r\n";
 		if (function_exists('wp_nonce_field')) {
 			wp_nonce_field('subscribe2-manage_subscribers' . $s2nonce);
 		}
+		echo "<h2>" . __('Add/Remove Subscribers', 'subscribe2') . "</h2>\r\n";
 		echo "<p>" . __('Enter addresses, one per line or comma-separated', 'subscribe2') . "<br />\r\n";
 		echo "<textarea rows=\"2\" cols=\"80\" name=\"addresses\"></textarea></p>\r\n";
 		echo "<input type=\"hidden\" name=\"s2_admin\" />\r\n";
@@ -1292,7 +1306,7 @@ class s2class {
 			echo "<input type=\"submit\" name=\"csv\" value=\"" . __('Save Emails to CSV File', 'subscribe2') . "\" class=\"button\" /></td></tr></table>\r\n";
 			if ($reminderform) {
 				echo "<input type=\"hidden\" name=\"reminderemails\" value=\"" . $reminderemails . "\" />\r\n";
-				echo "<p class=\"submit\" align=\"right\" style=\"border-top:none;\"><input type=\"submit\" name=\"reminder\" value=\"" . __('Send Reminder Email', 'subscribe2') . "\" /></p>\r\n";
+				echo "<p class=\"submit\" align=\"right\" style=\"border-top:none;\"><input type=\"submit\" name=\"remind\" value=\"" . __('Send Reminder Email', 'subscribe2') . "\" /></p>\r\n";
 			}
 			echo "<table width=\"100%\"><tr><td valign=\"bottom\">" . $strip . "</td>\r\n";
 			echo "<td align=\"right\"><p class=\"submit\" align=\"right\" style=\"border-top: none;\"><input type=\"submit\" name=\"process\" value=\"" . __('Process', 'subscribe2') . "\" /></p>\r\n";
@@ -1301,7 +1315,7 @@ class s2class {
 
 		echo "<table cellpadding=\"2\" cellspacing=\"2\" width=\"100%\">";
 		if (!empty($subscribers)) {
-			$subscriber_chunks = array_chunk($subscribers, 25);
+			$subscriber_chunks = array_chunk($subscribers, $this->subscribe2_options['entries']);
 			$chunk = $page - 1;
 			$subscribers = $subscriber_chunks[$chunk];
 			echo "<tr class=\"$alternate\">\r\n";
@@ -1394,9 +1408,12 @@ class s2class {
 				$this->reset();
 				echo "<div id=\"message\" class=\"updated fade\"><p><strong>$this->options_reset</strong></p></div>";
 			} elseif ($_POST['submit']) {
-
 				// BCClimit
-				$this->subscribe2_options['bcclimit'] = $_POST['bcc'];
+				if ( (is_numeric($_POST['bcc'])) && ($_POST['bcc'] >= 0) ) {
+					$this->subscribe2_options['bcclimit'] = $_POST['bcc'];
+				}
+				// admin_email
+				$this->subscribe2_options['admin_email'] = $_POST['admin_email'];
 
 				// send as author or admin?
 				$sender = 'author';
@@ -1443,8 +1460,26 @@ class s2class {
 				$this->subscribe2_options['confirm_email'] = $_POST['confirm_email'];
 				$this->subscribe2_options['remind_email'] = $_POST['remind_email'];
 
+				// excluded categories
+				if (!empty($_POST['category'])) {
+					$exclude_cats = implode(',', $_POST['category']);
+				} else {
+					$exclude_cats = '';
+				}
+				$this->subscribe2_options['exclude'] = $exclude_cats;
+				// allow override?
+				(isset($_POST['reg_override'])) ? $override = '1' : $override = '0';
+				$this->subscribe2_options['reg_override'] = $override;
+
 				// default WordPress page where Subscribe2 token is placed
-				$this->subscribe2_options['s2page'] = $_POST['page'];
+				if ( (is_numeric($_POST['page'])) && ($_POST['page'] >= 0) ) {
+					$this->subscribe2_options['s2page'] = $_POST['page'];
+				}
+
+				// Number of subscriber per page
+				if ( (is_numeric($_POST['entries'])) && ($_POST['entries'] > 0) ) {
+					$this->subscribe2_options['entries'] =$_POST['entries'];
+				}
 
 				// show meta link?
 				($_POST['show_meta'] == '1') ? $showmeta = '1' : $showmeta = '0';
@@ -1457,17 +1492,6 @@ class s2class {
 				// show widget in Presentation->Widgets
 				($_POST['widget'] == '1') ? $showwidget = '1' : $showwidget = '0';
 				$this->subscribe2_options['widget'] = $showwidget;
-
-				// excluded categories
-				if (!empty($_POST['category'])) {
-					$exclude_cats = implode(',', $_POST['category']);
-				} else {
-					$exclude_cats = '';
-				}
-				$this->subscribe2_options['exclude'] = $exclude_cats;
-				// allow override?
-				(isset($_POST['reg_override'])) ? $override = '1' : $override = '0';
-				$this->subscribe2_options['reg_override'] = $override;
 
 				//automatic subscription
 				$this->subscribe2_options['autosub'] = $_POST['autosub'];
@@ -1485,6 +1509,7 @@ class s2class {
 		}
 		// show our form
 		echo "<div class=\"wrap\">";
+		echo "<h2>" . __('Subscribe2 Settings', 'subscribe2') . "</h2>\r\n";
 		echo "<form method=\"post\" action=\"\">\r\n";
 		if (function_exists('wp_nonce_field')) {
 			wp_nonce_field('subscribe2-options_subscribers' . $s2nonce);
@@ -1492,17 +1517,40 @@ class s2class {
 		echo "<input type=\"hidden\" name=\"s2_admin\" value=\"options\" />\r\n";
 		echo "<input type=\"hidden\" id=\"jsbcc\" value=\"" . $this->subscribe2_options['bcclimit'] . "\" />";
 		echo "<input type=\"hidden\" id=\"jspage\" value=\"" . $this->subscribe2_options['s2page'] . "\" />";
-		
+		echo "<input type=\"hidden\" id=\"jsentries\" value=\"" . $this->subscribe2_options['entries'] . "\" />";
+
 		// settings for outgoing emails
-		echo "<h2>" . __('Delivery Options', 'subscribe2') . ":</h2>\r\n";
+		echo "<h2>" . __('Notification Settings', 'subscribe2') . "</h2>\r\n";
 		echo "<p>";
 		echo __('Restrict the number of recpients per email to (0 for unlimited)', 'subscribe2') . ': ';
 		echo "<span id=\"s2bcc_1\"><span id=\"s2bcc\" style=\"background-color: #FFFBCC\">" . $this->subscribe2_options['bcclimit'] . "</span> ";
-		echo "<a href=\"#\" onclick=\"s2_show('bcc')\">" . __('Edit', 'subscribe2') . "</a></span>\n";
+		echo "<a href=\"#\" onclick=\"s2_show('bcc'); return false;\">" . __('Edit', 'subscribe2') . "</a></span>\n";
 		echo "<span id=\"s2bcc_2\">\r\n";
 		echo "<input type=\"text\" name=\"bcc\" value=\"" . $this->subscribe2_options['bcclimit'] . "\" size=\"3\" />\r\n";
-		echo "<a href=\"#\" onclick=\"s2_update('bcc');\">". __('Update', 'subscribe2') . "</a>\n";
-		echo "<a href=\"#\" onclick=\"s2_revert('bcc');\">". __('Revert', 'subscribe2') . "</a></span>\n";
+		echo "<a href=\"#\" onclick=\"s2_update('bcc'); return false;\">". __('Update', 'subscribe2') . "</a>\n";
+		echo "<a href=\"#\" onclick=\"s2_revert('bcc'); return false;\">". __('Revert', 'subscribe2') . "</a></span>\n";
+
+		echo "<br /><br />" . __('Send Admins notifications for new', 'subscribe2') . ': ';
+		echo "<input type=\"radio\" name=\"admin_email\" value=\"subs\"";
+		if ('subs' == $this->subscribe2_options['admin_email']) {
+			echo " checked=\"checked\"";
+		}
+		echo " /> " . __('Subscriptions', 'subscribe2') . " &nbsp;&nbsp;";
+		echo "<input type=\"radio\" name=\"admin_email\" value=\"unsubs\"";
+		if ('unsubs' == $this->subscribe2_options['admin_email']) {
+			echo " checked=\"checked\"";
+		}
+		echo " /> " . __('Unsubscriptions', 'subscribe2') . " &nbsp;&nbsp;";
+		echo "<input type=\"radio\" name=\"admin_email\" value=\"both\"";
+		if ('both' == $this->subscribe2_options['admin_email']) {
+			echo " checked=\"checked\"";
+		}
+		echo " /> " . __('Both', 'subscribe2') . " &nbsp;&nbsp;";
+		echo "<input type=\"radio\" name=\"admin_email\" value=\"none\"";
+		if ('none' == $this->subscribe2_options['admin_email']) {
+			echo " checked=\"checked\"";
+		}
+		echo " /> " . __('Neither', 'subscribe2');
 
 		echo "<br /><br />" . __('Send Emails for Pages', 'subscribe2') . ': ';
 		echo "<input type=\"radio\" name=\"pages\" value=\"yes\"";
@@ -1585,18 +1633,39 @@ class s2class {
 		echo "<textarea rows=\"9\" cols=\"60\" name=\"remind_email\">" . stripslashes($this->subscribe2_options['remind_email']) . "</textarea><br /><br />\r\n";
 		echo "</td></tr></table><br />\r\n";
 
-		// Main options
-		echo "<h2>" . __('Appearance Options', 'subscribe2') . "</h2>\r\n";
+		// excluded categories
+		echo "<h2>" . __('Excluded Categories', 'subscribe2') . "</h2>\r\n";
+		echo"<p>";
+		echo "<strong><em style=\"color: red\">" . __('Posts assigned to any Excluded Category do not generate notifications and are not included in digest notifications', 'subscribe2') . "</em></strong><br />\r\n";
+		echo"</p>";
+		$this->display_category_form(explode(',', $this->subscribe2_options['exclude']));
+		echo "<center><input type=\"checkbox\" name=\"reg_override\" value=\"1\"";
+		if ('1' == $this->subscribe2_options['reg_override']) {
+			echo " checked=\"checked\"";
+		}
+		echo " /> " . __('Allow registered users to subscribe to excluded categories?', 'subscribe2') . "</center><br />\r\n";
+		
+		// Appearance options
+		echo "<h2>" . __('Appearance', 'subscribe2') . "</h2>\r\n";
 		echo"<p>";
 		
 		// WordPress page ID where subscribe2 token is used
 		echo __('Set default Subscribe2 page as ID', 'subscribe2') . ': ';
 		echo "<span id=\"s2page_1\"><span id=\"s2page\" style=\"background-color: #FFFBCC\">" . $this->subscribe2_options['s2page'] . "</span> ";
-		echo "<a href=\"#\" onclick=\"s2_show('page')\">" . __('Edit', 'subscribe2') . "</a></span>\n";
+		echo "<a href=\"#\" onclick=\"s2_show('page'); return false;\">" . __('Edit', 'subscribe2') . "</a></span>\n";
 		echo "<span id=\"s2page_2\">\r\n";
 		echo "<input type=\"text\" name=\"page\" value=\"" . $this->subscribe2_options['s2page'] . "\" size=\"3\" />\r\n";
-		echo "<a href=\"#\" onclick=\"s2_update('page');\">". __('Update', 'subscribe2') . "</a>\n";
-		echo "<a href=\"#\" onclick=\"s2_revert('page');\">". __('Revert', 'subscribe2') . "</a></span>\n";
+		echo "<a href=\"#\" onclick=\"s2_update('page'); return false;\">". __('Update', 'subscribe2') . "</a>\n";
+		echo "<a href=\"#\" onclick=\"s2_revert('page'); return false;\">". __('Revert', 'subscribe2') . "</a></span>\n";
+
+		// Number of subscribers per page
+		echo "<br /><br />" . __('Set the number of Subscribers displayed per page', 'subscribe2') . ': ';
+		echo "<span id=\"s2entries_1\"><span id=\"s2entries\" style=\"background-color: #FFFBCC\">" . $this->subscribe2_options['entries'] . "</span> ";
+		echo "<a href=\"#\" onclick=\"s2_show('entries'); return false;\">" . __('Edit', 'subscribe2') . "</a></span>\n";
+		echo "<span id=\"s2entries_2\">\r\n";
+		echo "<input type=\"text\" name=\"entries\" value=\"" . $this->subscribe2_options['entries'] . "\" size=\"3\" />\r\n";
+		echo "<a href=\"#\" onclick=\"s2_update('entries'); return false;\">". __('Update', 'subscribe2') . "</a>\n";
+		echo "<a href=\"#\" onclick=\"s2_revert('entries'); return false;\">". __('Revert', 'subscribe2') . "</a></span>\n";
 
 		// show link to WordPress page in meta
 		echo "<br /><br /><input type=\"checkbox\" name=\"show_meta\" value=\"1\"";
@@ -1619,18 +1688,6 @@ class s2class {
 		}
 		echo " /> " . __('Enable Subscribe2 Widget?', 'subscribe2') . "<br /><br />\r\n";
 		echo"</p>";
-
-		// excluded categories
-		echo "<h2>" . __('Excluded Categories', 'subscribe2') . "</h2>\r\n";
-		echo"<p>";
-		echo "<strong><em style=\"color: red\">" . __('Posts assigned to any Excluded Category do not generate notifications and are not included in digest notifications', 'subscribe2') . "</em></strong><br />\r\n";
-		echo"</p>";
-		$this->display_category_form(explode(',', $this->subscribe2_options['exclude']));
-		echo "<center><input type=\"checkbox\" name=\"reg_override\" value=\"1\"";
-		if ('1' == $this->subscribe2_options['reg_override']) {
-			echo " checked=\"checked\"";
-		}
-		echo " /> " . __('Allow registered users to subscribe to excluded categories?', 'subscribe2') . "</center><br />\r\n";
 
 		//Auto Subscription for new registrations
 		echo "<h2>" . __('Auto Subscribe', 'subscribe2') . "</h2>\r\n";
@@ -1934,9 +1991,8 @@ class s2class {
 			echo "<div id=\"message\" class=\"updated\"><strong><p>" . $message . "</p></strong></div>\r\n";
 		}
 		// show our form
-		echo "<div class=\"wrap\"><h2>" . __('Send email to all subscribers', 'subscribe2') . "</h2>\r\n";
+		echo "<div class=\"wrap\"><h2>" . __('Send an email to subscribers', 'subscribe2') . "</h2>\r\n";
 		echo "<form method=\"post\" action=\"\">\r\n";
-		echo "<p>";
 		if (function_exists('wp_nonce_field')) {
 			wp_nonce_field('subscribe2-write_subscribers' . $s2nonce);
 		}
@@ -1945,7 +2001,7 @@ class s2class {
 		} else {
 			$subject = __('A message from ', 'subscribe2') . get_option('blogname');
 		}
-		echo __('Subject', 'subscribe2') . ": <input type=\"text\" size=\"69\" name=\"subject\" value=\"" . $subject . "\" /> <br /><br />";
+		echo "<p>" . __('Subject', 'subscribe2') . ": <input type=\"text\" size=\"69\" name=\"subject\" value=\"" . $subject . "\" /> <br /><br />";
 		echo "<textarea rows=\"12\" cols=\"75\" name=\"content\">" . $body . "</textarea>";
 		echo "<br /><br />\r\n";
 		echo __('Recipients: ', 'subscribe2');
@@ -2181,7 +2237,7 @@ class s2class {
 	Display our form; also handles (un)subscribe requests
 	*/
 	function filter($content = '') {
-		if ( ('' == $content) || (1 == $this->filtered) || (!strstr($content, '<!--subscribe2-->')) ) { return $content; }
+		if ( ('' == $content) || (!strstr($content, '<!--subscribe2-->')) ) { return $content; }
 		$this->s2form = $this->form;
 
 		global $user_ID;
@@ -2217,7 +2273,7 @@ class s2class {
 							$this->add();
 							$status = $this->send_confirm('add');
 							// set a variable to denote that we've already run, and shouldn't run again
-							$this->filtered = 1; //set this to not send duplicate emails
+							$this->filtered = 1;
 							if ($status) {
 								$this->s2form = $this->confirmation_sent;
 							} else {
@@ -2278,7 +2334,6 @@ class s2class {
 	*/
 	function title_filter($title) {
 		// don't interfere if we've already done our thing
-		if (1 == $this->filtered) { return; }
 		if (in_the_loop()) {
 			return __('Subscription Confirmation', 'subscribe2');
 		} else {
@@ -2471,7 +2526,7 @@ class s2class {
 					$check = true;
 				}
 			}
-			//is the current post set by the user to
+			// is the current post set by the user to
 			// not generate a notification email?
 			$s2mail = get_post_meta($post->ID, 's2mail', true);
 			if (strtolower(trim($s2mail)) == 'no') {
@@ -2616,9 +2671,7 @@ class s2class {
 
 		// add actions for automatic subscription based on option settings
 		add_action('register_form', array(&$this, 'register_form'));
-		if ('yes' == $this->subscribe2_options['autosub']) {
-			add_action('user_register', array(&$this, 'register'));
-		}
+		add_action('user_register', array(&$this, 'register'));
 		if ('wpreg' == $this->subscribe2_options['autosub']) {
 			add_action('register_post', array(&$this, 'register_post'));
 		}
